requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.2
diff-match-patch==20230430
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-ipware==7.0.1
django-redis==5.4.0
django-rest-authemail==2.1.7
django-rq==2.10.2
djangorestframework==3.15.2
idna==3.9
python-dotenv==1.0.1
python-ipware==3.0.0
redis==5.0.8
requests==2.32.3
rq==1.16.2
sqlparse==0.5.1
tablib==3.5.0
urllib3==2.2.3
